+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Crash on Vulkan when blending 3D textures.
+- `RuntimeUtilities.DestroyVolume()` works as expected now.
 
 ### Changed
 - Improved volume texture blending.